vlog: Correct include paths for various modules
When using the XILINX ISE front-end, the include paths fail since there's no
way (or atleast I haven't figured a way) to pass it an include search path the
way you would do so in fuse with -i.

We correct the include paths such that it would work both with fuse and the ISE
front-end.

Signed-off-by: Joel Fernandes <joelf@ti.com>